websocket: fix IndexError when no subprotocols
It is possible to get an empty list of subprotocols from a client (e.g. from Chrome on Android). In this case, we get an indexError because event.subprotocols is a list with length 0.

This adds an extra check for the case of the empty list. We use the value `None` if the list is empty since `handshake.H1Handshake._accept` checks for this value later.